Addition and Deletion of nodes implemented
When nodes added and deleted , ids are in right order.
All the searching would fail accordingly.
Wrote methods to assign ids in order after operation.
This is used after deletion and it works.

Graphlib nodes were created with a different id. Fixed that bug to take ids
directly from the core tree datastructure.

Start commenting tree.js and organizing it for better understanding