Edit node, edge and add edge operations implemented
For editing the labels of node and edge, a message box is popped up.
User inputs the entry and updates it.

For add edge operation, two buttons are added to each and every button.
One to select itself as a parent and the other to add to under it as child.