Bug fix - Dealing with children more than 9
The numbering system for assigning Ids for each node is by using a numeric
represented by a character. At the max we could represented 10 childnodes
using 0 to 9. Hence the representation is using as x.y.z.a etc. where as x,y
are numbers that can be of any length